Fixed Linting moved CustomDomainLangSource to its own file.
