+++ src/wovnio/wovnphp/API.php
@@ -103,7 +103,6 @@ class API
             $default_lang = $store->settings['default_lang'];
             $defaultCustomDomainLang = $customDomainLangs->getCustomDomainLangByLang($default_lang);
             return CustomDomainLangUrlHandler::changeToNewCustomDomainLang($uri, $currentLangDomainLang, $defaultCustomDomainLang);
-
         } else {
             return $uri;
         }
